refactor(prefixmap)!: moved some errors to error.rs and changed errors export path
