+++ prefixmap/src/iri/deref.rs
@@ -1,35 +1,6 @@
-use crate::{PrefixMap, PrefixMapError};
-use iri_s::{IriS, IriSError};
-use thiserror::Error;
-
-use crate::IriRefError;
-
-/// Represents all the possible errors that can occur when dereferencing IRIs
-#[derive(Debug, Error, Clone)]
-pub enum DerefError {
-    /// An error originating from the [`iri_s`] crate
-    #[error(transparent)]
-    IriSError(#[from] IriSError),
-
-    /// An error originating when obtaining an IRI from a [`PrefixMap`]
-    ///
-    /// ## Fields
-    /// - `error`: The underlying [`PrefixMapError`]
-    #[error("Error obtaining IRI for '{alias}:{local}': {error}")]
-    DerefPrefixMapError {
-        alias: String,
-        local: String,
-        error: Box<PrefixMapError>,
-    },
-
-    /// An error occured when trying to dereference a prefixed name without a prefix map
-    #[error("No prefix map to dereference prefixed name {prefix}{local}")]
-    NoPrefixMapPrefixedName { prefix: String, local: String },
-
-    /// An error occured when trying to obtain an IRI from an [`IriRef`]
-    #[error(transparent)]
-    UnderefError(#[from] IriRefError),
-}
+use crate::error::DerefError;
+use crate::PrefixMap;
+use iri_s::IriS;
 
 /// A trait for dereferencing IRIs with an optional base IRI and a [`PrefixMap`]
 pub trait Deref {
